+++ CHANGELOG.md
@@ -11,4 +11,5 @@ Features:
 - Added unitTest for class Brownie\HttpClient\Exception\ValidateException.
 - Added unitTest for class Brownie\HttpClient\Request.
 - Added unitTest for class Brownie\HttpClient\Response.
+- Added unitTest for class Brownie\HttpClient\Headers.
 - Refactoring.